+++ src/main/java/com/tristankechlo/livingthings/init/ModFoods.java
@@ -11,5 +11,9 @@ public class ModFoods {
 	public static final Food BANANA = (new Food.Builder()).nutrition(4).saturationMod(0.3F).build();
 	public static final Food ELEPHANT = (new Food.Builder()).nutrition(4).saturationMod(0.6F).build();
 	public static final Food COOKED_ELEPHANT = (new Food.Builder()).nutrition(7).saturationMod(0.9F).build();
+	public static final Food LION = (new Food.Builder()).nutrition(3).saturationMod(0.5F).build();
+	public static final Food COOKED_LION = (new Food.Builder()).nutrition(6).saturationMod(0.8F).build();
+	public static final Food GIRAFFE = (new Food.Builder()).nutrition(3).saturationMod(0.5F).build();
+	public static final Food COOKED_GIRAFFE = (new Food.Builder()).nutrition(6).saturationMod(0.8F).build();
 
 }